fix: correct details for `list-full-paths` = false
When `list-full-paths` detail output from `lcov` won't include full
paths to files. That means that nothing will get matched in
`lineRefersToChangedFile()` because `lineWithFilename` includes
only basename, not dir.

Fix this by remembering the last dir when looping over output from
`lcov`. Then join the dir and the current file basename, only then
check if we have a match or not.